+++ src/part/mod.rs
@@ -1,6 +1,6 @@
 pub mod archive;
 pub mod fhs;
-pub mod pack;
+pub mod group;
 pub mod prune;
 pub mod store;
 pub mod version;